triage/triage_assistant/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pyyaml>=6.0
llama-cpp-python>=0.2.70
kokoro-onnx>=0.4.0
onnxruntime>=1.17.0
faster-whisper>=1.0.0
numpy>=1.24.0
soundfile>=0.12.0
python-multipart>=0.0.9
